Index implementation.
Read indexes from redis, updated indexes on insert, update and delete.

Initial implementation uses EVAL. This could be improved to use WATCH on
the index and normal commands which are repeated on failure.

Initial version works correctly, but is bandwidth heavy. It also suffers
from being non symmetrical in that the insert operation reads from the
local index list, while the erase operation reads from the remote index
list. They should in all cases be equal but this is not guaranteed.

Additionally, because the EVAL script accesses a dynamic key list (based
on indexed columns) it may not play nice with redis cluster.

All good reasons to refactor later with WATCH MULTI EVAL