Indexes working - not fully validated.
Indexes are now differentiated based on their type. It is possible to
provide optimised behaviour for indexes on integral and floating point
types, while falling back to suboptimal behaviour for string indexes.

Direct lookup on string indexes is fast (expected to be the common case)
while relative lookups (> >= < <=) at the moment fall back to a full
table scan.